Fixes to enable persistence layer.

+++ petstore-services/src/main/java/com/iwd/petstore_services/PetStoreServiceImpl.java
@@ -1,5 +1,6 @@
 package com.iwd.petstore_services;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -11,10 +12,7 @@ public class PetStoreServiceImpl implements PetStoreService {
 
 	private final PetRespository petRepository;
 
-	public PetStoreServiceImpl() {
-		petRepository = null;
-	}
-
+	@Autowired
 	public PetStoreServiceImpl(PetRespository aPetRepository) {
 		petRepository = aPetRepository;
 	}
@@ -37,6 +35,11 @@ public class PetStoreServiceImpl implements PetStoreService {
 
 		Pet petToBeDeleted = petRepository.findOne(petId);
 
-		petRepository.delete(petToBeDeleted);
+		if (petToBeDeleted!=null) {
+			petRepository.delete(petToBeDeleted);
+		}
+		else {
+			// TODO - put an error log here!
+		}
 	}
 }
